// iterpro-frontend/src/app/+state/session-analysis-store/ngrx/session-analysis-store.actions.enum.ts
export enum SessionAnalysisActions {
  ResetStore = '[SessionAnalysisStore] Reset Store',
  InitStore = '[SessionAnalysisStore] Init Store',
  InitStoreSuccess = '[SessionAnalysisStore] Init Store Success',
  initStoreFailure = '[SessionAnalysisStore] Init Store Failure',
  SetRouteParams = '[SessionAnalysisStore] Set Route Params',
  SelectView = '[SessionAnalysisStore] Select View',
  SelectSessionView = '[SessionAnalysisStore] Select Session Analysis View',
  SelectPeriodView = '[SessionAnalysisStore] Select Period Analysis View',
  LoadSeasonInfo = '[SessionAnalysisStore] Load Season Info',
  LoadSeasonInfoSuccess = '[SessionAnalysisStore] Load Season Info Success',
  LoadSeasonInfoFailure = '[SessionAnalysisStore] Load Season Info Failure',
  SelectSession = '[SessionAnalysisStore] Select Session',
  LoadSessions = '[SessionAnalysisStore] Load Sessions',
  LoadSessionsSuccess = '[SessionAnalysisStore] Load Sessions Success',
  LoadSessionsFailure = '[SessionAnalysisStore] Load Sessions Failure',
  LoadSessionDetails = '[SessionAnalysisStore] Load Session Details',
  LoadSessionDetailsSuccess = '[SessionAnalysisStore] Load Session Details Success',
  LoadSessionDetailsFailure = '[SessionAnalysisStore] Load Session Details Failure',
  LoadPeriodTotal = '[SessionAnalysisStore] Load Period Total Details',
  LoadPeriodTotalSuccess = '[SessionAnalysisStore] Load Period Total Details Success',
  LoadPeriodTotalFailure = '[SessionAnalysisStore] Load Period Total Details Failure',
  LoadPeriodTrend = '[SessionAnalysisStore] Load Period Trend Details',
  LoadPeriodTrendSuccess = '[SessionAnalysisStore] Load Period Trend Details Success',
  LoadPeriodTrendFailure = '[SessionAnalysisStore] Load Period Trend Details Failure',
  SelectSessionPlayers = '[SessionAnalysisStore] Select Session Team Players',
  SelectIndividualPlayer = '[SessionAnalysisStore] Select individual player',
  SelectPeriodPlayers = '[SessionAnalysisStore] Select Period players',
  SelectSessionSplit = '[SessionAnalysisStore] Select Session Team Split',
  SelectIndividualSplits = '[SessionAnalysisStore] Select Session Individual Split',
  SelectPeriodSplits = '[SessionAnalysisStore] Select Period Splits',
  SelectFilter = '[SessionAnalysisStore] Select Filter',
  SelectSessionType = '[SessionAnalysisStore] Select Session Type',
  SelectMetrics = '[SessionAnalysisStore] Select Metrics',
  SelectBubbleMetrics = '[SessionAnalysisStore] Select Bubble Metrics',
  SelectAdvancedOption = '[SessionAnalysisStore] Select Advanced Option',
  LoadAdvancedSessionData = '[SessionAnalysisStore] Load Advanced Session Data',
  LoadAdvancedPeriodData = '[SessionAnalysisStore] Load Advanced Period Data',
  LoadAdvancedDataSuccess = '[SessionAnalysisStore] Load Advanced Data Success',
  LoadAdvancedDataFailure = '[SessionAnalysisStore] Load Advanced Data Failure',
  SelectDatePeriod = '[SessionAnalysisStore] Select Date Period',
  ToggleSidebar = '[SessionAnalysisStore] Toggle Sidebar',
  ToggleThresholds = '[SessionAnalysisStore] Toggle Thesholds',
  TogglePercentage = '[SessionAnalysisStore] Toggle Percentage',
  ToggleOrder = '[SessionAnalysisStore] Toggle Order',
  ToggleLabels = '[SessionAnalysisStore] Toggle Labels',
  ToggleBubble = '[SessionAnalysisStore] Toggle Bubble',
  DownloadPDF = '[SessionAnalysisStore] Download PDF',
  DownloadSessionTeamCSV = '[SessionAnalysisStore] Download Session Team CSV',
  DownloadPeriodCSV = '[SessionAnalysisStore] Download Period CSV',
  DownloadFileSuccess = '[SessionAnalysisStore] Download Session CSV',
  GoCalendarSession = '[SessionAnalysisStore] Go To Calendar Session',
  ShowUploadDialog = '[SessionAnalysisStore] Show Upload Dialog',
  HideUploadDialog = '[SessionAnalysisStore] Hide Upload Dialog',
  SaveAttachments = '[SessionAnalysisStore] Save Attachments',
  SaveAttachmentsSuccess = '[SessionAnalysisStore] Save Attachments Completed',
  SaveAttachmentsFailed = '[SessionAnalysisStore] Save Attachments Failed',
}